requirements.txt
fuzzywuzzy==0.18.0
mysql-connector==2.2.9
numpy==1.18.5
pandas==1.0.4
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2020.1
six==1.15.0
tqdm==4.46.1
